Remove old logic and make refactor the default
